+++ ubo_app/utils/types.py
@@ -1,0 +1,6 @@
+"""Some useful type definitions."""
+
+from collections.abc import Callable
+from typing import TypeAlias
+
+Subscriptions: TypeAlias = list[Callable[[], None]]